feat: basic terraform setup for digital ocean
Provides a basic Terraform setup to create a bunch of droplets on Digital Ocean.

The genesis node and the other nodes are differentiated using tags. This is going to be necessary
for using Ansible, so we can have one playbook target the genesis node and another target all the
others.